feat: fixed some naming issues
